# improve cat performance on CPU