--- a/aten/src/ATen/native/cpu/CatKernel.h
+++ b/aten/src/ATen/native/cpu/CatKernel.h
@@ -3,10 +3,11 @@
 #include <ATen/core/Tensor.h>
 #include <ATen/native/DispatchStub.h>
 #include <ATen/core/ITensorListRef.h>
+#include <ATen/TensorGeometry.h>
 
 namespace at { namespace native {
 
-using cat_serial_fn = void(*)(const Tensor &, const MaterializedITensorListRef&, int64_t);
-DECLARE_DISPATCH(cat_serial_fn, cat_serial_stub);
+using cat_contig_fn = void(*)(const Tensor &, const MaterializedITensorListRef&, int64_t, bool);
+DECLARE_DISPATCH(cat_contig_fn, cat_contig_stub);
 
 }}  // namespace at::native